requirements.txt
numpy
opencv-python
torch
torchvision
matplotlib
scikit-image
pycryptodome
reedsolo
flask
requests
